% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_functions.R
\name{alpha_cronbach}
\alias{alpha_cronbach}
\title{Compute Cronbach's alpha}
\usage{
alpha_cronbach(S)
}
\arguments{
\item{S}{Covariance matrix generated by \code{var()}.}
}
\value{
The return value is a number (double): the estimated value of Cronbach's 
alpha.
}
\description{
\code{alpha_cronbach()} takes a covariance matrix of data, \code{S}, 
that has been created from a matrix in which each column represents a 
variable. It returns Cronbach's alpha for the battery. It is not 
exported, but it is called by \linkInt{reliability}.
}
\author{
Joseph F. Lucke
}
