% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bullock-package.R
\docType{package}
\name{Bullock-package}
\alias{Bullock-package}
\alias{_PACKAGE}
\alias{Bullock}
\title{Bullock: convenience functions and tools for table generation}
\description{
This package provides tools that I use in my code. Most of the 
tools are simple convenience functions. Three of the 
functions---\linkInt{regTable}, \linkInt{latexTable}, and 
\linkInt{latexTablePDF}---are more elaborate tools for making tables, and 
they are designed to be used together. See the  
\href{../doc/tables.html}{Building better tables in less time} vignette for
an overview.
}
\details{
Of the convenience functions, I most often use \linkInt{qw}, 
\linkInt{reliability}, \linkInt{meanNA}, and \linkInt{sumNA}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbullock35/Bullock}
  \item Report bugs at \url{https://github.com/jbullock35/Bullock/issues}
}

}
\author{
\strong{Maintainer}: John Bullock \email{john@johnbullock.org}

}
\keyword{internal}
