% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTablePDF.R
\name{latexTablePDF}
\alias{latexTablePDF}
\title{Renders a LaTeX table as a PDF file.}
\usage{
latexTablePDF(
  latexTable,
  container = TRUE,
  containerFilename = NULL,
  outputFilenameStem = "latexTable",
  writePDF = TRUE,
  writeTex = FALSE,
  overwriteExisting = FALSE,
  verbose = FALSE,
  continuedFloat = FALSE,
  continuedFloatStar = FALSE,
  firstPageEmpty = TRUE,
  firstTableNumber = NULL,
  openPDFOnExit = TRUE
)
}
\arguments{
\item{latexTable}{Object of class \code{latexTable} (typically produced by 
\code{latexTable()} or a list of such objects.}

\item{container}{Logical variable. Should the LaTeX code in \code{latexTable}
be inserted into a LaTeX file that contains a complete LaTeX preamble and
the \code{\\begin{document}} and \code{\\end{document}} tags? If you want 
\code{latexTablePDF} to produce a PDF file, \code{container} must be 
\code{TRUE}. But if you just want to write \code{latexTable} to disk as 
a set of .tex files that you will insert into your own larger LaTeX 
document, \code{container} should probably be \code{FALSE}.}

\item{containerFilename}{A string. Specifies the path of the "container" 
LaTeX file into which the \code{latexTable} table or tables will be 
inserted to make a complete LaTeX file that can be rendered as PDF.}

\item{outputFilenameStem}{A string. It is the path and name of the file
is to be saved to disk, up to the extension. For example, if you want to  
save "myTable.pdf" to disk, set \code{outputFilenameStem = "mytable"}.}

\item{writePDF}{Logical variable. Should a PDF file be saved to disk?}

\item{writeTex}{Logical variable. Should a .tex file be saved to disk?}

\item{overwriteExisting}{Logical variable. Should files to be saved overwrite
existing files that have the same names?}

\item{verbose}{Logical variable. \code{latexTablePDF} calls
\code{pdflatex} to render PDF files, and if \code{verbose} is \code{TRUE}, 
all of the output from \code{pdflatex} will be printed to screen.\cr\cr\cr\cr}

\item{continuedFloat}{Logical variable. Should be \code{TRUE} if the table 
or tables to be rendered are part of a series and should all share the 
same number. For example, \code{continuedFloat} should be true if 
\code{latexTable} is a list of tables and you all want them to be 
numbered as Table{\NB}3.}

\item{continuedFloatStar}{Logical variable. Should be \code{TRUE} if the table 
or tables to be rendered are part of a series, and if all tables should 
share the same number but be distinguished by some secondary character.
For example, \code{continuedFloatStar} should be true if \code{latexTable} 
is a list of tables and you all want them to be numbered as Table~3a, 
Table~3b, etc.\cr
  \indent \code{continuedFloat} and \code{continuedFloatStar} cannot both 
be \code{TRUE}.}

\item{firstPageEmpty}{Logical variable. If \code{TRUE}, the page that 
contains the first table in \code{latexTable} will have an empty header
and footer. (Technically, \code{latexTablePDF} will insert 
\code{\\thispagestyle{empty}} into the code block that contains the first
table in \code{latexTable}.)}

\item{firstTableNumber}{Integer. What number should the first table in 
  \code{latexTable} have?\cr
  \indent By default, the table numbering will be "natural." For example, 
if \code{latexTablePDF} is used to produce a PDF document, the first table 
in the document will be Table{\NB}1. And if \code{latexTablePDF} is instead
used to create a .tex file that is to be included in a larger LaTeX 
document, the number of the first \code{latexTable} table will be determined
by the number of preceding tables in the document. For example, if 
you use \code{latexTablePDF} to create a .tex file that you then insert 
into a larger LaTeX document, and if the .tex file is preceded in the 
LaTeX document by two tables, the first table in \code{latexTable} will be 
Table{\NB}3.\cr\cr\cr\cr}

\item{openPDFOnExit}{Logical variable. Open the PDF file after it is created
by \code{latexTablePDF}? This argument has an effect only on Windows, and
only if \code{writePDF} is \code{TRUE}.}
}
\description{
\code{latexTablePDF} takes an object produced by \code{latexTable} and 
writes a PDF file. It can also write the corresponding .tex file.
}
\details{
Although \code{latexTablePDF} produces PDF files by default, it is also 
useful for creating .tex files. For example, you may have a list of tables
produced by \code{latexTable()} and a complex LaTeX 
document that contains many different sections and tables. When 
\code{latexTablePDF} is used with \code{writeTex = TRUE}, it will produce a 
single file that contains LaTeX code for all of the tables in your list. 
You can insert those tables into your LaTeX document by adding a 
single \code{\\input} or \code{\\include} command to your LaTeX document.
Doing so will either insert the tables directly (if you ran  
\code{latexTable()} with \code{callCommand = FALSE} or make available to  
you a set of LaTeX macros that you can use to insert the tables into  
arbitrary places in your LaTeX document.  [PUT THIS INTO THE VIGNETTE!]
}
\note{
\emph{Required LaTeX tools.} If \code{writePDF} is \code{TRUE}, 
  \code{pdflatex} must be installed on your system. (It is part of almost 
  every LaTeX installation.) In addition, if \code{writePDF} is \code{TRUE}  
  and \code{containerFilename} is \code{NULL} (the default), the following 
  LaTeX packages must be installed: \code{array, booktabs, caption, 
  fancyhdr, geometry, pdflscape, } and \code{ragged2e}. Most of these, and 
  perhaps all of them, are already part of your LaTeX installation.\cr\cr\cr\cr
}
\examples{
\dontrun{
  data(iris)
  lm1 <- lm(Sepal.Length ~ Petal.Length,               data = iris)
  lm2 <- lm(Sepal.Length ~ Petal.Length + Petal.Width, data = iris)
  lm3 <- lm(Sepal.Length ~ Petal.Length * Petal.Width, data = iris)
  lmList <- list(lm1, lm2, lm3)
  rT1 <- regTable(lmList)
  lT1 <- latexTable(
    mat        = rT1,
    colNames   = qw("(1) (2) (3)"),
    rowNames   = c("Intercept", "Petal length", "Petal width", "Petal length $\\\\times$ petal width"),
    footerRows = list(
      c("Number of observations", sapply(lmList, nobs)),
      c("R$^2$", sapply(lmList, function (x) round(summary(x)$r.squared, 2)))
    ),
    spacerColumns      = c(0, 2, 4),
    spacerColumnsWidth = '.67em',
    spacerRows         = 1,  # separate Intercept row from other rows
    caption            = '\\\\textit{Sepal length as a function of petal length and petal width.} Lorem ipsum dolor.'
  )
  latexTablePDF(lT1)
}
}
\seealso{
Other functions for making tables: 
\code{\link{latexTable}()},
\code{\link{regTable}()}
}
\concept{functions for making tables}
