% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{lt_colNames_default}
\alias{lt_colNames_default}
\title{Compute default column names in calls to latexTable().}
\usage{
lt_colNames_default(
  mat = parent.frame()$mat,
  SE_table = parent.frame()$SE_table
)
}
\arguments{
\item{mat}{A matrix, typically a \code{regTable} object.}

\item{SE_table}{Logical variable. See \code{\link{latexTable}}.}
}
\value{
A vector of strings. Each string is a column{\NB}name.
}
\description{
If \code{colnames(mat)} is not \code{NULL}, this function will use 
\code{colnames(mat)} as the \code{colNames} argument in \code{latexTable()}.
If \code{colnames(mat)} is \code{NULL}, column names will be determined by
\code{\link{lt_colNumbers}()}.
}
\details{
The function is not exported and is intended to be called only by 
\code{latexTable()}.
}
