% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{lt_colNumbers}
\alias{lt_colNumbers}
\title{Automatically determine column names of the form (1), (2), etc.}
\usage{
lt_colNumbers(mat = sys.frame(-1)$mat, SE_table = sys.frame(-1)$SE_table)
}
\arguments{
\item{mat}{A matrix, typically a \code{regTable} object.}

\item{SE_table}{Logical variable. See \code{\link{latexTable}}.}
}
\value{
A vector of strings. If \code{SE_table} is \code{TRUE}, the vector 
elements are "(1)", "(2)", etc., where the last column number is 
\code{ncol(mat)/2}. If \code{SE_table} is \code{FALSE}, the vector 
elements are "(1)", "(2)", etc., where the last column number is simply 
\code{ncol(mat)}.
}
\description{
Given \code{mat} and \code{SE_table}, this function determines appropriate
column-number names of the form "(1)", "(2)", etc.
}
