% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{lt_footer}
\alias{lt_footer}
\title{Compute default footers for latexTable() objects.}
\usage{
lt_footer(
  mat = parent.frame()$mat,
  rowNames = parent.frame()$rowNames,
  SE_table = parent.frame()$SE_table,
  decimalPlaces = parent.frame()$decimalPlaces
)
}
\arguments{
\item{mat}{A matrix, typically a \code{regTable} object.}

\item{rowNames}{Character vector. See \linkInt{latexTable}.}

\item{SE_table}{Logical variable. See \linkInt{latexTable}.}
}
\value{
A list of string vectors. Each list element contains the strings
needed to produce a footer{\NB}row.
}
\description{
The default footer row or rows are determined in this way: if 
\code{SE_table} is \code{FALSE} or \code{rowNames} is \code{NULL}, 
no footer rows are produced. Otherwise, a footer row will be added for each 
of the following attributes of \code{mat}: "r.squared", "SER", and "N". If 
\code{mat} lacks one of those attributes, there will be no corresponding 
footer{\NB}row.   
The function is not exported and is intended to be called only by 
\linkInt{latexTable}.
}
