% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{lt_nobsRow}
\alias{lt_nobsRow}
\title{Specify a footer row that indicates the number of observations for each regression.}
\usage{
lt_nobsRow(mat = parent.frame()$mat)
}
\arguments{
\item{mat}{A \code{regTable} object.}
}
\value{
A vector of strings. The first element in the vector is "Number of 
observations". The remaining elements are the numbers of observations for 
each regression in \code{mat}.
}
\description{
Given a \code{mat} produced by \code{\link{regTable}()}, this function 
returns a footer row that indicates the number of observations for each 
model in \code{mat}.
}
