% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{lt_rSquaredRow}
\alias{lt_rSquaredRow}
\title{Specify a footer row that indicates \ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}} for each regression.}
\usage{
lt_rSquaredRow(
  mat = parent.frame()$mat,
  decimalPlaces = parent.frame()$decimalPlaces
)
}
\arguments{
\item{mat}{A matrix, typically a \code{regTable} object.}

\item{decimalPlaces}{Integer. See \code{\link{latexTable}}.}
}
\value{
A vector of strings. The first element in the vector is "R$^2$". 
The remaining elements are strings that indicate 
\ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}} for each model in \code{mat}.
The strings are rounded to the number of digits specified by the 
\code{decimalPlaces} argument.
}
\description{
Given a \code{mat} produced by \code{\link{regTable}()} in which all
regressions are of class \code{lm}, this function returns a footer row that 
indicates \ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}} for each model in \code{mat}.
}
