% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{lt_SER_row}
\alias{lt_SER_row}
\title{Specify a footer row that indicates the standard error of regression for each model}
\usage{
lt_SER_row(
  mat = sys.frame(-1)$mat,
  decimalPlaces = sys.frame(-1)$decimalPlaces
)
}
\arguments{
\item{mat}{A matrix, typically a \code{regTable} object.}

\item{decimalPlaces}{Integer. See \code{\link{latexTable}}.}
}
\value{
A vector of strings. The first element in the vector is "Std. error
of regression". The remaining elements are strings that indicate the SER 
for each model in \code{mat}.  The strings are rounded to the number of 
digits specified by the \code{decimalPlaces} argument.
}
\description{
Given a \code{mat} produced by \code{\link{regTable}()} in which all
regressions are of class \code{lm}, this function returns a footer row that 
indicates the standard error of regression (i.e., \eqn{\sigma}, the 
"residual standard error") for each model in \code{mat}.
}
