% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePercentage.R
\name{makePercentage}
\alias{makePercentage}
\title{Convert a number or a string to a percentage.}
\usage{
makePercentage(x, dp = 0)
}
\arguments{
\item{x}{Object of class "character," "integer," or "numeric."}

\item{dp}{Integer specifying the number of decimal places in the returned string.}
}
\value{
A string, e.g., "57%"
}
\description{
Takes an object \code{x} of class "character" or "numeric". Returns an 
object of class "character" in which the number is expressed as a 
percentage.
}
\details{
By default, the returned string will have no decimal digits: it will be
"57%", not "57.0%". But if \code{dp} is greater than 0, the returned  
string will always have \code{dp} decimal places.
}
\examples{
makePercentage(.5)            # "50\%"
makePercentage(.555)          # "56\%"
makePercentage(.555, dp = 4)  # "55.5000\%"
makePercentage(5)             # "500\%"
makePercentage(5, dp = 2)     # "500.00\%"

makePercentage(1:3)           # "100\%" "200\%" "300\%"

}
