% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value_functions.R
\name{missingValueFunctions}
\alias{missingValueFunctions}
\alias{lNA}
\alias{lNAv}
\alias{meanNA}
\alias{sdNA}
\alias{sumNA}
\alias{varNA}
\title{Missing-value helper functions.}
\usage{
lNA(x, verbose = FALSE)

lNAv(x)

meanNA(x)

sdNA(x)

sumNA(x)

varNA(x)
}
\arguments{
\item{x}{An R object.}

\item{verbose}{Logical variable. If \code{TRUE}, \code{lNA} will print the 
lengths of \code{x} before and after \code{NA} values have been removed.}
}
\description{
Functions to make code a little clearer. These are mainly ordinary functions,
like \code{mean()}, with \code{na.rm} set to \code{TRUE}.
}
\details{
\code{lNA(x)} returns its value silently. \code{lNAv} is shorthand for 
\code{lNA(x, verbose = TRUE)}; it returns the same value as \code{lNA(x)}  
but also prints the lengths of the vector before and after NAs are removed.
}
\examples{
x <- c(1:3, NA, 5)
lNA(x)
lNAv(x)

sum(x)    # returns NA
sumNA(x)  # returns 11
meanNA(x)

sdNA(x)
varNA(x)
}
