\name{modal_value}
\alias{modal_value}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Find modal value of a vector}
\description{Find modal value of a vector.}
\usage{modal_value(x, na.rm = FALSE)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a vector}
  \item{na.rm}{Remove NAs before modal value is determined}
}
\details{}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{Ken Williams.  Function copied from \url{http://stackoverflow.com/a/8189441/697473}.}

\note{If there are multiple values, this function will return only the first.}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
