% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalValue.R
\name{modalValue}
\alias{modalValue}
\title{Find modal value of a vector.}
\usage{
modalValue(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector.}

\item{na.rm}{Logical variable.}
}
\description{
Find modal value of a vector. If there are multiple modal values, all will
be returned.
}
\details{
Before this package was released, it returned only the first mode if there
were multiple modes. It now returns all modes. See the examples.
}
\examples{
modalValue(qw("a b b"))      # [1] "b"
modalValue(qw("a a b b c"))  # [1] "a" "b"
modalValue(1:3)              # [1] 1 2 3

}
\author{
Ken Williams. See \url{http://stackoverflow.com/a/8189441/697473}.
}
