% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveToDF.R
\name{moveToDF}
\alias{moveToDF}
\title{Move or copy "freestanding" variables into a data frame.}
\usage{
moveToDF(pattern = NULL, move = TRUE)
}
\arguments{
\item{pattern}{String that specifies a regular expression. It is
\code{NULL} by default, in which case all variables in the environment are
examined for inclusion in the data frame.}

\item{move}{Logical variable.}
}
\value{
Data frame containing all one-dimensional variables that have names 
matching \code{pattern} and that have the modal length of those variables.
}
\description{
Variables are specified by \code{pattern}, which is a regular expression.
The modal length of all variables in the calling environment that match
\code{pattern} is determined. Matching variables are then moved to a data 
frame (or copied to a data frame if \code{move} is \code{FALSE}). 
\code{pattern} may be \code{NULL}, in which case all variables in the 
calling environment will be examined.
  Variables that have the "dim" attribute -- for example, arrays and 
matrices -- will not be moved into the new data frame. Functions will 
never be moved into the new data frame, either.
}
