% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNAmatrix.R
\name{noNAmatrix}
\alias{noNAmatrix}
\title{Perform listwise deletion on a matrix.}
\usage{
noNAmatrix(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
\code{noNAmatrix} performs "listwise deletion" on a matrix, removing all 
rows that contain any missing (NA) values.
}
\details{
This function is deprecated. Use \code{na.omit} instead.
}
\examples{
noNAmatrix(matrix(c(1:8, NA), nrow=3))
}
