% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF_crop.R
\name{PDF_crop}
\alias{PDF_crop}
\title{Crop white margins of a PDF file.}
\usage{
PDF_crop(
  input,
  dirOutput = NULL,
  outputExtension = "_crop",
  deleteOriginal = TRUE,
  openCropped = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{input}{A string. Path and name of file to be cropped. If \code{input}
does not end with ".pdf", ".pdf" will be appended to it. This is the only 
required argument.}

\item{dirOutput}{A string. Path of directory into which cropped file should 
be placed.}

\item{outputExtension}{A string. The new file will have the same name as
the old file, except that \code{outputExtension} will be inserted just 
before the ".pdf" extension.}

\item{deleteOriginal}{Logical, defaulting to \code{TRUE}. Should the
original file be deleted?}

\item{openCropped}{Logical, defaulting to \code{FALSE}. After the new file
is produced, should it be opened in your default PDF viewer? Ignored on 
non-Windows, non-Macintosh systems.}

\item{verbose}{Logical, defaulting to \code{FALSE}. If it is \code{TRUE},
two lines of information from \code{pdfcrop.exe} will be printed. They will
show your version of \code{pdfcrop.exe}, the number of pages in the file
that it produced, and the path and filename of the file that it produced.}
}
\description{
\code{PDF_crop()} removes ("crops") the white margins around a PDF. It 
requires that pdfcrop.exe be installed and in your path. To get pdfcrop.exe,
install MikTeX or Tex Live, or just use \code{tinytex::tlmgr_install('pdfcrop')}.
}
\examples{
tmpFile <- tempfile(fileext = ".pdf")
pdf(tmpFile)
  plot(1:10, 1:10)
dev.off()
PDF_crop(tmpFile)
}
\seealso{
\code{\link[knitr:plot_crop]{knitr::plot_crop()}}, which is less
flexible in its handling of PDF files but can handle other file formats.
}
