% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAll.R
\name{printAll}
\alias{printAll}
\title{Print all rows of a tibble.}
\usage{
printAll(x)
}
\arguments{
\item{x}{Object, perhaps of class "tbl_df"}
}
\description{
By default, tibbles---the default data frames created by tidyverse 
packages---print only as many rows as your console window can display.
\code{printAll()} instead prints all of the tibble's rows. If it is
called on an object other than a tibble, it just calls \code{print()}.
}
\examples{
x <- tibble::tibble(a = 1:100, b = 101:200)
printAll(x)

}
