% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regTable.R
\name{regTable}
\alias{regTable}
\title{Create a matrix of regression output from a list of regression models.}
\usage{
regTable(
  objList,
  colNames = NULL,
  rowsToRemove = NULL,
  rowsToKeep = NULL,
  clusterVar = NULL
)
}
\arguments{
\item{objList}{list of regression objects. This is the only required 
argument. \code{regTable()} has been tested with objects of classes 
\code{ivreg, glm, lm,} and \code{plm,} and with the objects produced by the 
\code{estimatr} package. It should work with other regression objects, too 
-- that is, with any regression objects for which the \code{coef()} and
\code{vcov()} functions can be used.}

\item{colNames}{A vector of strings as long as \code{length(objList)}.}

\item{rowsToRemove}{A vector of strings, which may specify regular 
expressions. Variables in the regressions whose names match the strings will 
be omitted from the \code{regTable} output. This argument overrides 
\code{rowsToKeep.}}

\item{rowsToKeep}{A vector of strings, which may specify regular 
expressions. Variables in the regressions whose names match the strings will 
be kept in the \code{regTable} output. All other variables will be omitted.
If \code{rowsToRemove} is specified, this argument has no effect.}

\item{clusterVar}{Either \code{NULL}, a list of length{\NB}1, or a list of 
length \code{length(objList)}. If \code{NULL} (the default), 
\code{regTable()} will report the standard errors indicated by your 
regression objects. Otherwise, the argument to \code{clusterVar} should 
indicate the clusters for the corresponding regression objects in
\code{objList}. Clustered SEs for "lm" objects will be produced by 
\code{multiwayvcov::cluster.vcov}, and clustered SEs for "ivreg" objects 
will be produced by \code{ivpack::cluster.robust.se}. \code{regTable()} 
does not support clustering for other kinds of regression objects.}
}
\value{
A matrix in which the columns are estimates and 
standard errors -- two columns for each model. The matrix has an "N"
attribute that indicates the number of observations for each regression. If
all regressions were of class \code{lm} (but not also class \code{glm}), it 
also has the "r.squared" and "SER" attributes. (The "SER" attribute 
indicates the standard error of regression -- AKA \eqn{\sigma} or the 
"residual standard error" --- for each model.)
}
\description{
\code{regTable()} takes a list of regression objects, such as those created 
by \code{lm()}. It returns a matrix in which the columns are estimates and 
standard errors -- two columns for each model.
}
\note{
Before \code{regTable()} was incorporated into this package,
it used the \code{rowsToKeep} argument differently: variables were kept 
only if the \emph{beginnings} of their names matched the strings in 
\code{rowsToKeep}.
}
\examples{
data(iris)
lm1 <- lm(Sepal.Length ~ Petal.Length,               data = iris)
lm2 <- lm(Sepal.Length ~ Petal.Length + Petal.Width, data = iris)
regTable(list(lm1, lm2))
regTable(list(lm1, lm2), colNames = c("Sepal length", "Sepal width"))
regTable(list(lm1, lm2), rowsToKeep = 'Length')
regTable(list(lm1, lm2), rowsToKeep = c('Intercept', 'Length'))
regTable(list(lm1, lm2), clusterVar = list(iris$Species))
}
\seealso{
Other functions for making tables: \linkInt{latexTable}, 
\linkInt{latexTablePDF}. See also the \href{../doc/tables.html}{Building better tables in less time} 
vignette.
}
