% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_functions.R
\name{reliability}
\alias{reliability}
\title{Compute and report Cronbach's alpha}
\usage{
reliability(x, ...)
}
\arguments{
\item{x}{Matrix of numbers}

\item{...}{additional arguments passed to \code{var()}}
}
\value{
List with two elements. The first element, \code{alpha}, is the 
estimated Cronbach's alpha for the battery of \code{n} variables. The 
second element, \code{alpha.if.deleted}, reports the estimated 
reliability of each combination of \code{n - 1} variables.
}
\description{
Given a matrix \code{x} in which every column represents a variable,
this function reports (a) the estimated reliability for the battery of 
\code{n} variables, and (b) the estimated variability for each possible 
combination of \code{n - 1} variables.
}
\examples{
data(iris)
reliability(iris[, 1:4])

}
\author{
Peter Ellis
}
