% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector to have a specified minimum and maximum.}
\usage{
rescale(x, newRange = c(0, 1))
}
\arguments{
\item{x}{Numeric object.}

\item{newRange}{Numeric vector of length{\NB}2.}
}
\description{
Rescale a vector to have a specified minimum and maximum.
}
\examples{
vec <- 1:10
rescale(vec, c(2, 5))

}
\seealso{
\linkInt{scale} and \code{\link[scales:rescale]{scales::rescale()}}
}
\author{
Simon D. Jackman
}
