% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakeToCamel.R
\name{snakeToCamel}
\alias{snakeToCamel}
\title{Transform a string from snake_case to camelCase.}
\usage{
snakeToCamel(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Given a string like "mean_attitude", the function returns "meanAttitude".
But for any given string segment involving an underscore character, the 
the segment will be transformed only if the character before the 
underscore is a lower-case letter. For example, "MEAN_attitude" won't be 
transformed. Nor will "PID3_early."
}
\examples{
snakeToCamel("party_ID")      # [1] "partyID"
snakeToCamel("hours_9_to_5")  # [1] "hours9_to_5" 

}
\seealso{
\linkInt{spaceToCamel}
}
