% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{spacerColumns_default}
\alias{spacerColumns_default}
\title{Compute default positions of spacer columns in calls to latexTable().}
\usage{
spacerColumns_default(
  mat = parent.frame()$mat,
  SE_table = parent.frame()$SE_table,
  rowNames = parent.frame()$rowNames
)
}
\arguments{
\item{mat}{Matrix.}

\item{SE_table}{Logical variable.}

\item{rowNames}{Vector.}
}
\value{
A vector of integers.
}
\description{
\code{spacerColumns_default} specifies the default \code{spacerColumns}
argument in calls to \code{latexTable}. It takes the values of \code{mat},
\code{SE_table}, and \code{rowNames} that are passed to \code{latexTable}.
From these values, it computes the default positions of spacer columns:
\itemize{
\item If \code{SE_table} is \code{FALSE}, there is a spacer column between every
column in \code{mat}.
\item If \code{SE_table} is \code{TRUE}, there is a spacer column after every
even-numbered column in \code{mat}, except for the last column.
\item If \code{rowNames} is not \code{NULL}, a spacer column is inserted between
the table's row names and the first column of data.\improveCSS
}
}
\details{
The function is not exported and is intended to be called only by
\code{latexTable()}.
}
