% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceToCamel.R
\name{spaceToCamel}
\alias{spaceToCamel}
\title{Transform a string that has spaces to camelCase.}
\usage{
spaceToCamel(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Given a string like "mean attitude", the function returns "meanAttitude".
But if the character just before the space is uppercase, the space will 
be replaced by an underscore (_): "TV news" will become "TV_news".
}
\examples{
spaceToCamel("colleges and universities")  # [1] "collegesAndUniversities"
spaceToCamel("TV news")                    # [1] "TV_news" 
spaceToCamel("TV News")                    # [1] "TV_News"

}
\seealso{
\linkInt{snakeToCamel}
}
