% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip0.R
\name{strip0}
\alias{strip0}
\title{Strip "0" from beginning of numbers printed by R.}
\usage{
strip0(x, dp = 2)
}
\arguments{
\item{x}{Object of class "character" or "numeric."}

\item{dp}{Integer specifying the number of decimal places in the returned string. Can be NULL, in which case no rounding is performed.}
}
\description{
Takes an object \code{x} of class "character" or "numeric". Returns an 
object of class "character" in which the leading zero has been removed. 
By default, it also rounds the number-string that it returns to two 
decimal places. Example: both 0.3445 and "0.3445" are transformed to 
".34".
}
\details{
If \code{x} has no leading zero, the returned object is still rounded to
\code{dp} decimal places.
}
\examples{
strip0("0.346")  #  ".35"
strip0(0.346)    #  ".35"
strip0(.346)     #  ".35"
strip0(5.346)    # "5.35"

strip0(c("0.789", ".2346", 53.3, 53.346))  # c(".79", ".23", "53.3", "53.35")

}
