% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexTable.R
\name{update.latexTable}
\alias{update.latexTable}
\title{Update a latexTable object with new arguments.}
\usage{
\method{update}{latexTable}(object, ...)
}
\arguments{
\item{object}{A \code{latexTable} object}

\item{...}{Arguments to \code{latexTable()}, e.g., \code{colNames}, 
\code{caption}.}
}
\value{
A \code{latexTable} object.
}
\description{
Each \code{latexTable} object stores, as an attribute, the call that 
produced{\NB}it. \code{update.latexTable()} updates the call by replacing 
arguments or adding new ones. It then calls \code{\link{latexTable}()} to produce
a new \code{latexTable} object.
}
\details{
\code{update.latexTable()} is adapted from \code{stats::update.default()}.  
It is a method for the generic \code{update()}.
}
\examples{
lT1 <- latexTable(matrix(1:16, nrow = 4))
lT2 <- update(lT1, mat = matrix(2:17, nrow = 8), commandName = "intTable")
}
