\name{varNA}
\alias{varNA}
\title{Calculate variance of vector after omitting NA values}
\description{Calculate variance of vector after omitting NA values}
\usage{varNA(x)}
\arguments{
  \item{x}{numeric vector, matrix, or data frame}
}
\details{The definition of \code{varNA} is \code{function(x) {var(x, na.rm = TRUE)}}.}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{\code{\link{var}}}
\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
