\docType{package}
\name{Bullock-package}
\alias{Bullock-package}
\alias{_PACKAGE}
\alias{Bullock}
\title{Bullock: convenience functions and tools for table generation}
\description{
This package provides tools that I use in my code. Most of the 
tools are simple convenience functions. Three of the 
functions---\linkInt{regTable}, \linkInt{latexTable}, and 
\linkInt{latexTablePDF}---are more elaborate tools for making tables, and 
they are designed to be used together. See the  
\href{../doc/tables.html}{Building better tables in less time} vignette for
an overview.
}
\details{
Of the convenience functions, I most often use \linkInt{qw}, 
\linkInt{reliability}, \linkInt{meanNA}, and \linkInt{sumNA}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbullock35/Bullock}
  \item Report bugs at \url{https://github.com/jbullock35/Bullock/issues}
}

}
\author{
  \itemize{
    \item Maintainer: John Bullock (\email{john@johnbullock.org})
    \item Florent Delmotte [contributor]
    \item Dirk Edelbuettel [contributor]
    \item Peter Ellis [contributor]
    \item Simon Jackman [contributor]
    \item JD Long [contributor]
    \item Joseph Lucke [contributor]
    \item Matt Pettis [contributor]
    \item Ken Williams [contributor]
  }
}

\keyword{internal}
