% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{.ls.objects}
\alias{.ls.objects}
\alias{lsos}
\title{List objects and their characteristics}
\usage{
.ls.objects(
  pos = 1,
  envir = as.environment(pos),
  pattern,
  order.by,
  decreasing = FALSE,
  n = NULL
)

lsos(..., n = 8)
}
\arguments{
\item{pos}{Numeric. Specifies the position, in the search list, of the
environment to search. Can be specified instead of \code{envir}.}

\item{envir}{Environment to search. Can be specified instead of \code{pos}.}

\item{pattern}{String. An optional regular expression; if it is specified,
only objects whose names match the regular expression will be returned.}

\item{order.by}{String, with value "Class", "Size", "Rows", or "Columns."
The returned data frame will be sorted by the specified column.}

\item{decreasing}{Logical variable. Should results be listed in decreasing
order of the \code{order.by} column? In \code{.ls.objects()} the default is
\code{FALSE}; in \code{lsos()}, it is \code{TRUE}.}

\item{n}{Numeric. Number of objects to list. In \code{lsos()}, the default 
is{\NB}8.}

\item{...}{Additional arguments to \code{.ls.objects()}.}
}
\value{
Data frame with columns "Class," "Size," "Row," and "Columns." The row 
names of the data frame are the names of objects in the environment.
}
\description{
\code{lsos()} and \code{ls.objects()} are like \code{ls()} in that they list  
the objects in a given environment. But they provide more detail: they show 
the class of each object, the amount of memory devoted to each objects, and
the number of rows or columns of each object (if applicable).
}
\details{
\code{lsos()} is a user-friendly wrapper for \code{.ls.objects()}. It is 
shorthand for \code{.ls.objects(order.by="Size", decreasing=TRUE, head=TRUE, n=8)}.

Both functions were created by Dirk Edelbuettel and modified by JD Long and
John Bullock. See \url{http://stackoverflow.com/questions/1358003/} for 
details.
}
\examples{
lsos()
lsos(pattern = '\\\\.df$')  # list only objects ending in ".df"
}
\author{
Both functions were created by Dirk Edelbuettel and modified by JD Long. 
See \url{http://stackoverflow.com/questions/1358003/} for details.
}
