% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regTable.R
\name{print.regTable}
\alias{print.regTable}
\title{Print regTable objects.}
\usage{
\method{print}{regTable}(x, decimalPlaces = getOption("Bullock.print.regTable.dp"), ...)
}
\arguments{
\item{x}{regTable object.}

\item{decimalPlaces}{Integer. Table entries will be rounded to this number
of digits after the decimal point. By default, it is the value of
\code{getOption("Bullock.print.regTable.dp")}, which is set to{\NB}2 when
the Bullock package is loaded. You can change the default by running
\code{options("Bullock.print.regTable.dp") <- N} at any time, where
\code{N} is your preferred number of digits after the decimal place.}

\item{...}{Arguments passed to \code{print.table()}.}
}
\description{
\code{print.regTable()} is the default \code{print} method for objects
produced by \linkInt{regTable}. The output that it produces differs from
\code{print.table()} output because \code{print.regTable()}
\itemize{
\item right-aligns row names;
\item inserts one space between each estimate and SE column, but two spaces
between estimate-SE column tiers, thereby making it easier to distinguish
between results from different regressions;
\item uses two rows of column headings: one to show the outcome for each
regression, and another to distinguish columns of coefficient estimates
from columns of SE estimates;
\item does not print attributes (e.g.,  "N", "r.squared").
}
}
