% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_functions.R
\name{reliability}
\alias{reliability}
\title{Compute and report Cronbach's alpha}
\usage{
reliability(x, ...)
}
\arguments{
\item{x}{Matrix of numbers}

\item{...}{additional arguments passed to \code{var()}}
}
\value{
The function prints output to the screen and invisibly returns a list with five elements:
\itemize{
\item \code{nrowBeforeListwiseDeletion} is the number of rows of \code{x}.
\item \code{nrowAfterListwiseDeletion} is the number of rows of \code{x} after listwise deletion.
\item \code{alpha} is Cronbach's alpha for \code{x} after listwise deletion.
\item \code{alphaSmall} is a vector of \code{ncol(x)} numbers. It shows what Cronbach's alpha is when we remove one variable at a time from \code{na.omit(x)}.
\item \code{alphaSmallAfterListwiseDeletion} is a vector of \code{ncol(x)} numbers. It shows what Cronbach's alpha is when we remove one variable at a time from \code{x} and then use listwise deletion on the new matrix.
}
}
\description{
Given a matrix \code{x} in which every column represents a variable,
this function reports (a) the estimated reliability (Cronbach's alpha)
for the battery of \code{n} variables, and (b) the estimated reliability
for each possible combination of \code{n - 1} variables. If \code{x}
contains missing values, listwise deletion will be done via
\code{na.omit(x)} so that alpha can be calculated.
}
\examples{
data(iris)
reliability(iris[, 1:4])

}
\author{
Peter Ellis

John G. Bullock
}
